+++ libraries/nestjs-libraries/src/media/media-provider.registry.spec.ts
@@ -61,6 +61,7 @@ describe('MediaModule', () => {
       'tavus',
       'togetherai',
       'vertex',
+      'wan',
     ]);
   });
 });